Fix: circular import
